refactor: [FE] Game 페이지 컴포넌트 리팩터링
- Intent
게임페이지와 결과페이지는 SketchbookSection 부분의 UI만 달라서 나머지부분을 재사용하려고 리팩터링 했어요.
